// gdr-web-version-mdp/gdr-web-version-mdp/src/app/@data/model/usuarioGdr.ts
export class UsuarioGdr {
  gestorEntidadId: number;
  entidadId: number;
  personaId: number;
  nombres: string;
  apePaterno: string;
  apeMaterno: string;
  nombreCompleto: string;
  unidadOrganicaId: number;
  unidadOrganica: string;
  detalleUoId: number;
  puesto: string;
  puestoId: number;
  fechaInicio: Date;
  fechaFin: Date;
  nombreArchivo: string;
  uuId: string;
  flagAsignacion: string;
  estadoRegistro: string;
  usuarioId: number;
  usuarioRolId: number;
  nroDocumento: number;
  colorEstado: string;
  estado: string;
  correo: string;
  disabledNotification: boolean;
  estadovalue: boolean;
}